Confirm to revoke token and delete provider

+++ src/components/providers/EditProvider.vue
@@ -6,7 +6,10 @@
     "urlProvider": "Configuration URL of the provider",
     "newClientId": "Generate a new client ID",
     "edit": "Confirm",
-    "remove": "Remove"
+    "remove": "Remove",
+    "warningremove": "Are you sure to remove this report provider ?",
+    "confirm": "Confirm",
+    "cancel": "Cancel"
 	},
 	"fr": {
 		"editprovider": "Edition d'un provider",
@@ -14,7 +17,10 @@
     "urlProvider": "URL de configuration",
     "newClientId": "Generer un nouveau client ID",
     "edit": "Confirmer",
-    "remove": "Supprimer"
+    "remove": "Supprimer",
+    "warningremove": "Etes-vous sûr de vouloir supprimer ce provider ?",
+    "confirm": "Confirmer",
+    "cancel": "Annuler"
 	}
 }
 </i18n>
@@ -99,30 +105,7 @@
         </div>
       </div>
       <div class="row">
-        <div class="offset-md-3 col-md-9 d-none d-sm-none d-md-block">
-          <button
-            type="submit"
-            class="btn btn-primary"
-            :disabled="loading"
-          >
-            {{ $t('edit') }}
-          </button>
-          <button
-            type="button"
-            class="btn btn-danger ml-3"
-            @click="deleteProvider"
-          >
-            {{ $t('remove') }}
-          </button>
-          <button
-            type="reset"
-            class="btn btn-secondary ml-3"
-            @click="cancel"
-          >
-            {{ $t('cancel') }}
-          </button>
-        </div>
-        <div class="col-12 d-md-none">
+        <div class="offset-md-3 col-12 col-sm-12 col-md-3">
           <button
             type="submit"
             class="btn btn-primary btn-block"
@@ -131,11 +114,47 @@
             {{ $t('edit') }}
           </button>
           <button
+            v-if="!confirmDelete"
             type="button"
             class="btn btn-danger btn-block"
             @click="deleteProvider"
           >
             {{ $t('remove') }}
+          </button>
+        </div>
+      </div>
+      <div
+        v-if="confirmDelete"
+        class="row mb-2"
+      >
+        <div class="offset-md-3 col-12 col-md-9">
+          <p
+            class="mt-2"
+          >
+            {{ $t('warningremove') }}
+          </p>
+        </div>
+      </div>
+      <div
+        v-if="confirmDelete"
+        class="row mb-2"
+      >
+        <div class="offset-md-3 col-12 col-sm-12 col-md-3">
+          <button
+            v-if="confirmDelete"
+            type="button"
+            class="btn btn-danger btn-block"
+            @click="deleteProvider"
+          >
+            {{ $t('confirm') }}
+          </button>
+          <button
+            v-if="confirmDelete"
+            type="button"
+            class="btn btn-secondary btn-block"
+            @click="confirmDelete=false"
+          >
+            {{ $t('cancel') }}
           </button>
         </div>
       </div>
@@ -166,7 +185,8 @@ export default {
 			newClientId: false,
 			show: false,
 			checkURL: false,
-			loading: false
+			loading: false,
+			confirmDelete: false
 		}
 	},
 	computed: {
@@ -228,16 +248,20 @@ export default {
 			this.$emit('done')
 		},
 		deleteProvider () {
-			this.$store.dispatch('deleteProvider', { albumID: this.albumID, clientID: this.clientID }).then(res => {
-				if (res.status !== 204) {
-					this.$snotify.error('Sorry, an error occured')
-				} else {
-					this.$snotify.success('Provider remove')
-					this.$emit('done')
-				}
-			}).catch(err => {
-				console.log(err)
-			})
+			if (!this.confirmDelete) {
+				this.confirmDelete = true
+			} else {
+				this.$store.dispatch('deleteProvider', { albumID: this.albumID, clientID: this.clientID }).then(res => {
+					if (res.status !== 204) {
+						this.$snotify.error('Sorry, an error occured')
+					} else {
+						this.$snotify.success('Provider remove')
+						this.$emit('done')
+					}
+				}).catch(err => {
+					console.log(err)
+				})
+			}
 		}
 	}
 }
